+++ Source/JxInternal.h
@@ -33,6 +33,7 @@ Copyright (c) 2016 James Boer
 #include "JxLogging.h"
 #include "JxCommon.h"
 #include "JxUnicode.h"
+#include "JxUnicodeCaseFolding.h"
 #include "JxGuid.h"
 #include "JxConversion.h"
 #include "JxSerialize.h"