+++ Source/JxScript.cpp
@@ -221,7 +221,7 @@ bool Script::Execute()
                     auto coll = var.GetCollectionItr().second;
                     if (itr != coll->end())
                         itr = coll->erase(itr);
-                    SetVariable(name, std::make_pair(itr, coll));
+                    SetVariableInternal(name, std::make_pair(itr, coll));
                 }
             }
             break;
@@ -684,7 +684,7 @@ bool Script::Execute()
 				String name;
 				m_execution.back().reader.Read(&name);
 				Variant val = Pop();
-				SetVariable(name, val);
+				SetVariableInternal(name, val);
 			}
 			break;
 			case Opcode::SetVarKey:
@@ -747,6 +747,11 @@ bool Script::Execute()
 
 Variant Script::GetVariable(const String & name) const
 {
+	return GetVariableInternal(FoldCase(name));
+}
+
+Variant Script::GetVariableInternal(const String & name) const
+{
 	auto & names = m_execution.back().names;
 	for (auto ritr = names.rbegin(); ritr != names.rend(); ++ritr)
 	{
@@ -789,6 +794,11 @@ void Script::Push(const Variant & value)
 
 void Script::SetVariable(const String & name, const Variant & value)
 {
+	SetVariableInternal(FoldCase(name), value);
+}
+
+void Script::SetVariableInternal(const String & name, const Variant & value)
+{
 	// Search down the variable stack for the variable
 	auto & names = m_execution.back().names;
 	for (auto ritr = names.rbegin(); ritr != names.rend(); ++ritr)
@@ -800,7 +810,7 @@ void Script::SetVariable(const String & name, const Variant & value)
 			if (index >= m_stack.size())
 			{
 				LogWriteLine("Attempted to access stack at invalid index");
-				return ;
+				return;
 			}
 			m_stack[itr->second] = value;
 			return;