+++ Source/JxUnicode.h
@@ -18,6 +18,12 @@ namespace Jinx
 	StringU16 ConvertUtf8ToUtf16(const String & utf8Str);
 	String ConvertUtf16ToUtf8(const StringU16 & utf16Str);
 	
+	bool IsCaseFolded(const String & source);
+	String FoldCase(const String & source);
+
+    void InitializeUnicode();
+    void ShutDownUnicode();
+
 };
 
 #endif // JX_UNICODE_H__