+++ Source/JxUnicodeCaseFolding.h
@@ -1,0 +1,28 @@
+/*
+The Jinx library is distributed under the MIT License (MIT)
+https://opensource.org/licenses/MIT
+See LICENSE.TXT or Jinx.h for license details.
+Copyright (c) 2016 James Boer
+*/
+
+#pragma once
+#ifndef JX_UNICODE_CASE_FOLDING_H__
+#define JX_UNICODE_CASE_FOLDING_H__
+
+
+namespace Jinx
+{
+
+    struct CaseFoldingData
+    {
+		char32_t sourceCodePoint;
+		char32_t destCodePoint1;
+		char32_t destCodePoint2;
+    };
+
+    size_t GetCaseFoldingTableSize();
+    const CaseFoldingData & GetCaseFoldingData(size_t index);
+
+};
+
+#endif // JX_UNICODE_CASE_FOLDING_H__